PP-624: move to hooks folder
