+++ src/hooks/index.ts
@@ -4,6 +4,7 @@ export { useNavigation } from './useNavigation'
 export { useRoute } from './useRoute'
 export { useThrottledEffect } from './useThrottledEffect'
 export { useValidatedState } from './useValidatedState'
+export { useExchangeRate } from './useExchangeRate'
 export { useMarketPrices } from './useMarketPrices'
 export { useToggleBoolean } from './useToggleBoolean'
 export { usePublicProfileNavigation } from './usePublicProfileNavigation'
