--- src/utils/market/index.ts
@@ -1,1 +1,0 @@
-export { useExchangeRate } from './useExchangeRate'
